server/managment/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
pytesseract==0.3.10
pytz==2023.3.post1
sqlparse==0.4.4
